rewrite `http` sector into client/server parts, clean up types, build helper setup for server
