+++ src/kernel_types.rs
@@ -453,9 +453,3 @@ pub fn en_wit_send_error_kind(kind: SendErrorKind) -> wit::SendErrorKind {
         SendErrorKind::Timeout => wit::SendErrorKind::Timeout,
     }
 }
-
-#[derive(Debug, Serialize, Deserialize)]
-pub enum MessageType {
-    Request,
-    Response,
-}
